+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 ### Added:
-- Added reward commands
+- 
 ### Fixed:
-- Fixed reward inventory
-- Fixed button position
+- Fixed double rewards
 ### Changed:
 -